Move to constants file and replace everywhere
