+++ docker/docker-tests.test.ts
@@ -3,6 +3,13 @@ import { ethers } from 'hardhat';
 import fs from 'fs';
 import path from 'path';
 import { AgglayerManager, AgglayerGER, AgglayerBridge, AggchainFEP, AgglayerGateway } from '../typechain-types';
+import {
+    DEFAULT_ADMIN_ROLE,
+    AGGCHAIN_DEFAULT_VKEY_ROLE,
+    AL_ADD_PP_ROUTE_ROLE,
+    AL_FREEZE_PP_ROUTE_ROLE,
+    AL_MULTISIG_ROLE,
+} from '../src/constants';
 
 const deployOutput = JSON.parse(fs.readFileSync(path.join(__dirname, './deploymentOutput/deploy_output.json'), 'utf8'));
 const {
@@ -82,11 +89,6 @@ describe('Docker build tests Contract', () => {
         const AgglayerGatewayFactory = await ethers.getContractFactory('AgglayerGateway');
         const AgglayerGatewayContract = AgglayerGatewayFactory.attach(aggLayerGatewayAddress) as AgglayerGateway;
         expect(AgglayerGatewayContract.target).to.equal(aggLayerGatewayAddress);
-        const DEFAULT_ADMIN_ROLE = ethers.ZeroHash;
-        const AGGCHAIN_DEFAULT_VKEY_ROLE = ethers.id('AGGCHAIN_DEFAULT_VKEY_ROLE');
-        const AL_ADD_PP_ROUTE_ROLE = ethers.id('AL_ADD_PP_ROUTE_ROLE');
-        const AL_FREEZE_PP_ROUTE_ROLE = ethers.id('AL_FREEZE_PP_ROUTE_ROLE');
-        const AL_MULTISIG_ROLE = ethers.id('AL_MULTISIG_ROLE');
         expect(await AgglayerGatewayContract.hasRole(DEFAULT_ADMIN_ROLE, admin)).to.be.equal(true);
         expect(await AgglayerGatewayContract.hasRole(AGGCHAIN_DEFAULT_VKEY_ROLE, admin)).to.be.equal(true);
         expect(await AgglayerGatewayContract.hasRole(AL_ADD_PP_ROUTE_ROLE, admin)).to.be.equal(true);
