+++ test/contractsv2/AggLayerGateway.test.ts
@@ -10,7 +10,7 @@ import {
     AL_ADD_PP_ROUTE_ROLE,
     AL_FREEZE_PP_ROUTE_ROLE,
     AL_MULTISIG_ROLE,
-} from './helpers/consts';
+} from '../../src/constants';
 
 describe('AgglayerGateway tests', () => {
     upgrades.silenceWarnings();
