+++ test/contractsv2/PolygonRollupManager.test.ts
@@ -33,7 +33,7 @@ import {
     STOP_EMERGENCY_ROLE,
     EMERGENCY_COUNCIL_ROLE,
     EMERGENCY_COUNCIL_ADMIN,
-} from './helpers/consts';
+} from '../../src/constants';
 
 type BatchDataStructEtrog = PolygonRollupBaseEtrog.BatchDataStruct;
