+++ test/contractsv2/PolygonRollupManagerAL-ECDSAMultisig.test.ts
@@ -23,12 +23,7 @@ import {
     computeSignersHash,
 } from '../../src/utils-common-aggchain';
 
-import { NO_ADDRESS } from '../../src/constants';
-import {
-    AGGCHAIN_DEFAULT_VKEY_ROLE,
-    AL_ADD_PP_ROUTE_ROLE,
-    AL_MULTISIG_ROLE,
-} from './helpers/consts';
+import { NO_ADDRESS, AGGCHAIN_DEFAULT_VKEY_ROLE, AL_ADD_PP_ROUTE_ROLE, AL_MULTISIG_ROLE } from '../../src/constants';
 
 describe('Polygon rollup manager aggregation layer v3: ECDSA Multisig', () => {
     // SIGNERS
