+++ test/contractsv2/PolygonRollupManagerUpgrade.test.ts
@@ -34,7 +34,7 @@ import {
     STOP_EMERGENCY_ROLE,
     EMERGENCY_COUNCIL_ROLE,
     EMERGENCY_COUNCIL_ADMIN,
-} from './helpers/consts';
+} from '../../src/constants';
 
 type BatchDataStructEtrog = PolygonRollupBaseEtrog.BatchDataStruct;
