+++ tools/initMigration/initMigration.ts
@@ -8,6 +8,7 @@ import { ethers } from 'hardhat';
 import { AgglayerManager } from '../../typechain-types';
 import { transactionTypes, genOperation } from '../utils';
 import initMigrationParams from './initMigration.json';
+import { UPDATE_ROLLUP_ROLE } from '../../src/constants';
 import { checkParams, getDeployerFromParameters, getProviderAdjustingMultiplierGas } from '../../src/utils';
 import { logger } from '../../src/logger';
 import { decodeScheduleData } from '../../upgrade/utils';
@@ -110,7 +111,6 @@ async function main() {
     } else {
         logger.info('Send tx to initMigration...');
         logger.info('Check deployer role');
-        const UPDATE_ROLLUP_ROLE = ethers.id('UPDATE_ROLLUP_ROLE');
         if ((await rollupManagerContract.hasRole(UPDATE_ROLLUP_ROLE, deployer.address)) === false) {
             logger.error(
                 'Deployer does not have admin role. Use the test flag on deploy_parameters if this is a test deployment',
